refactor: seperate getTags API fetcher
